tidy up and add ui element enable unittests
