+++ scripts/test/Engineering/texture/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_subdirectory(correction)
 add_subdirectory(polefigure)
 
-set(TEST_PY_FILES test_TextureUtils.py)
+set(TEST_PY_FILES test_TextureUtils.py test_xtal_helper.py)
 
 check_tests_valid(${CMAKE_CURRENT_SOURCE_DIR} ${TEST_PY_FILES})
